chore: use corresponding version for module
