+++ packages/dashpay-contract/schema/v1/dashpay.schema.json
