+++ packages/masternode-reward-shares-contract/schema/v1/masternode-reward-shares-documents.json
