+++ packages/masternode-reward-shares-contract/src/lib.rs
@@ -1,5 +1,5 @@
 mod error;
-pub mod v0;
+pub mod v1;
 
 pub use crate::error::Error;
 use platform_value::{Identifier, IdentifierBytes32};
@@ -31,7 +31,7 @@ pub fn load_definitions(platform_version: &PlatformVersion) -> Result<Option<Val
 }
 pub fn load_documents_schemas(platform_version: &PlatformVersion) -> Result<Value, Error> {
     match platform_version.system_data_contracts.withdrawals {
-        1 => v0::load_documents_schemas(),
+        1 => v1::load_documents_schemas(),
         version => Err(Error::UnknownVersionMismatch {
             method: "masternode_reward_shares_contract::load_documents_schemas".to_string(),
             known_versions: vec![1],
