+++ packages/rs-dpp/src/state_transition/state_transitions/identity/identity_credit_withdrawal_transition/mod.rs
@@ -18,7 +18,7 @@ use crate::state_transition::StateTransitionFieldTypes;
 use crate::identity::state_transition::OptionallyAssetLockProved;
 use crate::ProtocolError;
 use bincode::{Decode, Encode};
-use data_contracts::withdrawals_contract::v0::document_types::withdrawal::properties::OUTPUT_SCRIPT;
+use data_contracts::withdrawals_contract::v1::document_types::withdrawal::properties::OUTPUT_SCRIPT;
 use derive_more::From;
 use fields::*;
 use platform_serialization_derive::{PlatformDeserialize, PlatformSerialize, PlatformSignable};
