+++ packages/rs-drive-abci/src/execution/validation/state_transition/state_transitions/documents_batch/data_triggers/triggers/feature_flags/v0/mod.rs
@@ -10,7 +10,7 @@ use drive::state_transition_action::document::documents_batch::document_transiti
 use drive::state_transition_action::document::documents_batch::document_transition::document_create_transition_action::DocumentCreateTransitionActionAccessorsV0;
 use drive::state_transition_action::document::documents_batch::document_transition::DocumentTransitionAction;
 use dpp::system_data_contracts::feature_flags_contract;
-use dpp::system_data_contracts::feature_flags_contract::v0::document_types::update_consensus_params::properties
+use dpp::system_data_contracts::feature_flags_contract::v1::document_types::update_consensus_params::properties
 ::PROPERTY_ENABLE_AT_HEIGHT;
 use dpp::version::PlatformVersion;
