+++ packages/rs-drive-abci/src/execution/validation/state_transition/state_transitions/documents_batch/data_triggers/triggers/reward_share/v0/mod.rs
@@ -12,7 +12,7 @@ use dpp::document::DocumentV0Getters;
 use dpp::ProtocolError;
 use drive::state_transition_action::document::documents_batch::document_transition::document_base_transition_action::DocumentBaseTransitionActionAccessorsV0;
 use drive::state_transition_action::document::documents_batch::document_transition::document_create_transition_action::DocumentCreateTransitionActionAccessorsV0;
-use dpp::system_data_contracts::masternode_reward_shares_contract::v0::document_types::reward_share::properties
+use dpp::system_data_contracts::masternode_reward_shares_contract::v1::document_types::reward_share::properties
 ::{PAY_TO_ID, PERCENTAGE};
 use drive::drive::document::query::QueryDocumentsOutcomeV0Methods;
